pyproject.toml
[project]
name = "python-mcp"
version = "0.1.0"
description = "Python MCP Server for Code Graph Extraction"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "typing-extensions>=4.7.0",
    "python-dotenv>=0.20.0",
    "mcp"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["server", "agent", "code_grapher", "run_server"]
